トレーニング記録の表示をモーダルウィンドウからページに遷移する仕様に変更＋Xにシェアする機能を実装
