+++ frontend/src/components/Home/Body/TrainingRecord/TrainingTable/styles/training-record-container.css
@@ -7,7 +7,7 @@
 }
 
 /* タイトルのスタイル */
-.training-record-title {
+.training-record-titles {
   text-align: center;
 }
